Compiling, windows don't work

+++ carthage_update.sh
@@ -2,5 +2,12 @@
 
 carthage update --platform osx
 
-rm -rf ./src/Frameworks
-rsync -av --exclude='*.dSYM*' ./Carthage/Build/Mac/ ./src/Frameworks
+if [ $? -eq 0 ]
+then
+  rm -rf ./src/Frameworks
+  rsync -av --exclude='*.dSYM*' ./Carthage/Build/Mac/ ./src/Frameworks
+else
+  echo "Failed to build dependencies"
+  exit 1
+fi
+
